Template FormIntegrals over type
