+++ cpp/dolfinx/fem/assemble_vector_impl.h
@@ -164,7 +164,7 @@ void _lift_bc_cells(
 
   const std::function<void(T*, const T*, const T*, const double*, const int*,
                            const std::uint8_t*, const std::uint32_t)>& fn
-      = a.integrals().get_tabulate_tensor(FormIntegrals::Type::cell, 0);
+      = a.integrals().get_tabulate_tensor(IntegralType::cell, 0);
 
   // Prepare cell geometry
   const int gdim = mesh->geometry().dim();
@@ -286,7 +286,7 @@ void _lift_bc_exterior_facets(
 
   const std::function<void(T*, const T*, const T*, const double*, const int*,
                            const std::uint8_t*, const std::uint32_t)>& fn
-      = a.integrals().get_tabulate_tensor(FormIntegrals::Type::exterior_facet,
+      = a.integrals().get_tabulate_tensor(IntegralType::exterior_facet,
                                           0);
 
   // Prepare cell geometry
@@ -428,12 +428,12 @@ void assemble_vector(Eigen::Ref<Eigen::Matrix<T, Eigen::Dynamic, 1>> b,
   const Eigen::Array<T, Eigen::Dynamic, Eigen::Dynamic, Eigen::RowMajor> coeffs
       = pack_coefficients(L);
 
-  const FormIntegrals& integrals = L.integrals();
-  using type = fem::FormIntegrals::Type;
+  const FormIntegrals<PetscScalar>& integrals = L.integrals();
+  using type = fem::IntegralType;
   for (int i = 0; i < integrals.num_integrals(type::cell); ++i)
   {
     const auto& fn
-        = integrals.get_tabulate_tensor(FormIntegrals::Type::cell, i);
+        = integrals.get_tabulate_tensor(IntegralType::cell, i);
     const std::vector<std::int32_t>& active_cells
         = integrals.integral_domains(type::cell, i);
     fem::impl::assemble_cells(b, *mesh, active_cells, dofs, fn, coeffs,
@@ -443,7 +443,7 @@ void assemble_vector(Eigen::Ref<Eigen::Matrix<T, Eigen::Dynamic, 1>> b,
   for (int i = 0; i < integrals.num_integrals(type::exterior_facet); ++i)
   {
     const auto& fn
-        = integrals.get_tabulate_tensor(FormIntegrals::Type::exterior_facet, i);
+        = integrals.get_tabulate_tensor(IntegralType::exterior_facet, i);
     const std::vector<std::int32_t>& active_facets
         = integrals.integral_domains(type::exterior_facet, i);
     fem::impl::assemble_exterior_facets(b, *mesh, active_facets, *dofmap, fn,
@@ -454,7 +454,7 @@ void assemble_vector(Eigen::Ref<Eigen::Matrix<T, Eigen::Dynamic, 1>> b,
   {
     const std::vector<int> c_offsets = L.coefficients().offsets();
     const auto& fn
-        = integrals.get_tabulate_tensor(FormIntegrals::Type::interior_facet, i);
+        = integrals.get_tabulate_tensor(IntegralType::interior_facet, i);
     const std::vector<std::int32_t>& active_facets
         = integrals.integral_domains(type::interior_facet, i);
     fem::impl::assemble_interior_facets(b, *mesh, active_facets, *dofmap, fn,
@@ -752,9 +752,9 @@ void lift_bc(
     const std::vector<bool>& bc_markers1, double scale)
 {
   const Eigen::Matrix<T, Eigen::Dynamic, 1> x0(0);
-  if (a.integrals().num_integrals(fem::FormIntegrals::Type::cell) > 0)
+  if (a.integrals().num_integrals(fem::IntegralType::cell) > 0)
     _lift_bc_cells<T>(b, a, bc_values1, bc_markers1, x0, scale);
-  if (a.integrals().num_integrals(fem::FormIntegrals::Type::exterior_facet) > 0)
+  if (a.integrals().num_integrals(fem::IntegralType::exterior_facet) > 0)
     _lift_bc_exterior_facets<T>(b, a, bc_values1, bc_markers1, x0, scale);
 }
 //-----------------------------------------------------------------------------
@@ -766,9 +766,9 @@ void lift_bc(
     const Eigen::Ref<const Eigen::Matrix<T, Eigen::Dynamic, 1>>& x0,
     double scale)
 {
-  if (a.integrals().num_integrals(fem::FormIntegrals::Type::cell) > 0)
+  if (a.integrals().num_integrals(fem::IntegralType::cell) > 0)
     _lift_bc_cells(b, a, bc_values1, bc_markers1, x0, scale);
-  if (a.integrals().num_integrals(fem::FormIntegrals::Type::exterior_facet) > 0)
+  if (a.integrals().num_integrals(fem::IntegralType::exterior_facet) > 0)
     _lift_bc_exterior_facets(b, a, bc_values1, bc_markers1, x0, scale);
 }
 //-----------------------------------------------------------------------------
